2.3.16

+++ src/graph2cluster.h
@@ -6,8 +6,8 @@
 // node1 / node2 - 1-based
 // node2cluster - if node is unassigned node2cluster[node] will be -1
 // returns number of clusters created
-unsigned graph2cluster(const int *pnode1, const int *pnode2, const double *pweight, size_t num_edges,
-                       unsigned min_cluster_size, float cooling_rate, unsigned burn_in, unsigned *node2cluster, size_t num_nodes);
+unsigned graph2cluster(const int *pnode1, const int *pnode2, const double *pweight, uint64_t num_edges,
+                       unsigned min_cluster_size, float cooling_rate, unsigned burn_in, unsigned *node2cluster, uint64_t num_nodes);
 
 #endif
