+++ src/utils.cpp
@@ -16,12 +16,12 @@ SEXP tgs_finite(SEXP _x, SEXP _envir)
 {
 	try {
         TGStat tgstat(_envir);
-        size_t len = xlength(_x);
+        uint64_t len = xlength(_x);
 
         if (!isReal(_x) && !isInteger(_x))
             verror("\"x\" argument must be numeric or integer");
 
-        for (size_t i = 0; i < len; ++i) {
+        for (uint64_t i = 0; i < len; ++i) {
             if ((isReal(_x) && !R_FINITE(REAL(_x)[i])) || (isInteger(_x) && INTEGER(_x)[i] == NA_INTEGER))
                 rreturn(ScalarLogical(false));
         }
